Domains: Fix some usability issues with the custom nameservers controls
